pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claii"
version = "0.1.0"
description = "CLAII – a CLI-first AI coding agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[project.scripts]
claii = "claii.cli:main"
